requirements.txt
streamlit
pandas
websockets
freecurrencyapi
python-dotenv
openai